1.4.0
Обновление

+++ package/system/controllers/multilang/hooks/content_after_delete.php
@@ -4,13 +4,12 @@ class onMultilangContentAfterDelete extends cmsAction {
 
     public function run($data){
 		
-		list($ctype, $item) = $data;
-		
-		cmsCache::getInstance()->clean("multilang.multilang_con_{$data['ctype_name']}");
+		if ($this->model->db->isTableExists('multilang_con_' . $data['ctype_name'])){
+			cmsCache::getInstance()->clean("multilang.multilang_con_{$data['ctype_name']}");
+			$this->model->filterEqual('i.item_id', $data['item']['id'])->deleteFiltered('multilang_con_' . $data['ctype_name']);
+		}
 
-        $this->model->filterEqual('i.item_id', $data['item']['id'])->deleteFiltered('multilang_con_' . $data['ctype_name']);
-
-        return true;
+        return $data;
 
     }
